reorganized rust code into modules.
